Moved the apk tests to the alpine folder.
